Add more values and value constraints
- Make vm and compiler handle Nil and Booleans
- Make value literal f32 to save some memory space

+++ src/eval/interpreted_eval.rs
@@ -9,7 +9,7 @@ use std::fmt;
 
 #[derive(Debug, PartialEq)]
 pub enum EvalResult {
-    Number(f64),
+    Number(f32),
     String(String),
     Bool(bool),
     Nil,
